+++ Library/Vec3.h
@@ -6,33 +6,33 @@
 
 class Vec3 {
 private :
-    float vector[3];
+    double vector[3];
 
 public :
     Vec3() : vector{0.0,0.0,0.0} {}
-    Vec3(float v0, float v1, float v2) : vector{v0, v1, v2} {}
+    Vec3(double v0, double v1, double v2) : vector{v0, v1, v2} {}
     ~Vec3() = default;
 
     const Vec3& operator+() const { return *this; }
     Vec3 operator-() const { return Vec3(-vector[0], -vector[1], -vector[2]); }
-    float operator[](int i) const { return vector[i]; }
-    float& operator[](int i) { return vector[i]; };
+    double operator[](int i) const { return vector[i]; }
+    double& operator[](int i) { return vector[i]; };
 
-    float x() const {return vector[0];}
-    float y() const {return vector[1];}
-    float z() const {return vector[2];}
+    double x() const {return vector[0];}
+    double y() const {return vector[1];}
+    double z() const {return vector[2];}
     
     Vec3& operator+=(const Vec3 &v);
     Vec3& operator-=(const Vec3 &v);
     Vec3& operator*=(const Vec3 &v);
     Vec3& operator/=(const Vec3 &v);
-    Vec3& operator+=(const float d);
-    Vec3& operator-=(const float d);
-    Vec3& operator*=(const float d);
-    Vec3& operator/=(const float d);
+    Vec3& operator+=(const double d);
+    Vec3& operator-=(const double d);
+    Vec3& operator*=(const double d);
+    Vec3& operator/=(const double d);
 
-    float length();
-    float length_sqrd();
+    double length();
+    double length_sqrd();
 
     std::ostream& operator<<(std::ostream& out);
 
@@ -41,19 +41,19 @@ public :
     Vec3 operator*(const Vec3& v);
     Vec3 operator/(const Vec3& v);
 
-    float prod_escalar(const Vec3& v);
+    double prod_escalar(const Vec3& v);
     Vec3 prod_vec(const Vec3& v);
     Vec3 unit_vec();
 
 };
     // How deal with rho operations type 
-    // : float -> Vec -> Vec
-    Vec3 operator+(float d, const Vec3& v);
-    Vec3 operator+(const Vec3& v, float d);
-    Vec3 operator-(float d, const Vec3& v);
-    Vec3 operator-(const Vec3& v, float d);
-    Vec3 operator*(float d, const Vec3& v);
-    Vec3 operator*(const Vec3& v, float d);
-    Vec3 operator/(const Vec3& v, float d);
+    // : double -> Vec -> Vec
+    Vec3 operator+(double d, const Vec3& v);
+    Vec3 operator+(const Vec3& v, double d);
+    Vec3 operator-(double d, const Vec3& v);
+    Vec3 operator-(const Vec3& v, double d);
+    Vec3 operator*(double d, const Vec3& v);
+    Vec3 operator*(const Vec3& v, double d);
+    Vec3 operator/(const Vec3& v, double d);
 
 #endif
